requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
rvc3python>=0.6.0
machinevision-toolbox-python>=0.9.0
spatialmath-python>=1.1.0 